deprecation notice befor moving to v2.2
