+++ src/screen/main.hpp
@@ -1,7 +1,6 @@
 #ifndef MAIN_SCREEN_HPP
 #define MAIN_SCREEN_HPP
 
-#include "screen/entry.hpp"
 #include "game.hpp"
 #include "ftxui/component/screen_interactive.hpp"
